Fix double loading.
The change to `require_dependency` should fix this for most deployments
but also putting in a guard in cases were it still double loads.

I know it still double loads on a capistrano deployment (i.e. Ninefold)
due to capistrano using a shared vendor directory for all releases. The
plugin loads the file via the symlink while the Rails eager loading loads
via the load path (i.e. the shared vendor directory) so require_dependency
does not recognize it at the same file.