Force 3D secure to actually be used.
Although we were correctly generating the 3D secure token and we were passing
that info to the gateway, the gateway was actually preferring the non-secure
vault token instead.

This commit does a bit more override so that IF we are needing 3D secure it
completely replaces the `authorize` method rather than just decorating the
parameters. This replacement will ensure we use the secured token rather than
the vault non-secured vault token.