Remove teardown of device data collector
Often resource access has the following pattern:

* open resource
* do resource manipulation
* close resource

Working with files or database connections are good examples. Generally
the closing down of the resource access is necessary to ensure good
cleanup.

The Braintree device data API has the appearance of this pattern. There
is a setup and teardown method so the original code was implemented
under the idea of it working this way.

Unfortunately the teardown method really should not have been called
and it should only be called in rare scenarios were we need to reset
the device session.

Calling it right after capturing the device session had the effect of
not capturing device data and therefore the fraud prevention was
not effective.