Update code so that device_data is used when vaulting
The device_data was found to not actually be used. 1eadcc9 corrected how
the device_data was initialized in the client which fixed the problem
for when transacting the order but we got word that the device_data
still isn't being captured when the card is being vaulted.

This commit corrects this via several mechanisms:

* I found that the code which communicates the device data to the models
  works for the transaction but not for the vaulting because of the
  structure of the checkout process. To fix I created an additional
  decoration on  `update_from_params` which communicates the device
  data to the models on the payment stage. The necessitated adding
  `device_data` as a permitted param for source attributes.
* When the options are being built for vaulting the card the
  `options_for_payment` override was calling `super` twice rather than
  just merging both additional options in one line. This caused the
  first call (where the device data was merged in) to be ineffective.

In that last bullet point I also refactored the determination of the
merchant_account_id into being a simple method call instead of the
"presenter" object.

Finally there are two bits of code from 7daaded that seem to be wrong
that were corrected:

* options[:payment_method_nonce] assignment was changed from `true` to
  the encrypted data. In the happy case (encrypted data provided) this
  was equal but in the case where CSE isn't being used it caused a
  error in an unobvious place. Restored just using `true` to make the
  cause of the error more obvious.
* There was an attempt to save the merchant id for the current
  transaction in the global spree preferences. This didn't make sense
  to me (or Martin) so I have removed it.